pyproject.toml
[tool.poetry]
name = "learnpython-2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-login = "^0.6.3"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
flasgger = "^0.9.7.1"
werkzeug = "^3.0.4"
jwt = "^1.3.1"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
